refactor: move CurrentUserContext to 'contexts' folder
